Rename IT folders (#938)
(cherry picked from commit fb0d773f5c0fb0eba96e657dc8998e4fc7e54ca9)
