+++ .buildkite/it/serverless-pipeline.yml
@@ -1,5 +1,5 @@
 env:
-    RUN_FULL_CI_WHEN_CHANGED: pyproject.toml,.buildkite,it_serverless
+    RUN_FULL_CI_WHEN_CHANGED: pyproject.toml,.buildkite,it_tracks_serverless
 
 common:
   plugins:
