+++ it_tracks/test_all_tracks_and_challenges.py
